Documentation and code updates 1. LICENSE Terms changes 2. CONTRIBUTING.md documents changed 3. README modified for new code 4. Added Optional Arguments for Code 5. Shifted read.sh to scripts folder 6. Removed outdated drivers
