+++ drivers/README.md
@@ -1,0 +1,25 @@
+# Drivers for Selenium
+
+## About 
+* Selenium Python bindings provides a simple API to write functional/acceptance tests using Selenium WebDriver. Through Selenium Python API you can access all functionalities of Selenium WebDriver in an intuitive way.
+
+* Selenium Python bindings provide a convenient API to access Selenium WebDrivers like Firefox, Ie, Chrome, Remote etc. The current supported Python versions are 2.7, 3.5 and above.
+
+* There are two drivers that are officially supported for running code on this Repository. This application has been tested on both Chrome as well as Mozilla.
+
+## Downloading Python Bindings for Selenium
+Using pip, you can install selenium like this:
+
+```
+pip install selenium
+```
+
+You may consider using virtualenv to create isolated Python environments. Python 3.6 has pyvenv which is almost same as virtualenv.
+
+## Geckodriver (Mozilla)
+
+You can find the download link [here](https://github.com/mozilla/geckodriver/releases). Make sure you download the latest version of the Geckodriver if you have the latest updated Mozilla Version. 
+
+## Chrome Driver (Chrome)
+
+You can find the download link [here](https://sites.google.com/a/chromium.org/chromedriver/downloads). Make sure you download the latest version of the Geckodriver if you have the latest updated Chrome Version.
