+++ requirements.txt
@@ -1,2 +1,1 @@
-aiml==0.8.6
-selenium==3.4.3
+selenium==3.10.0
