--- wa-web/bin/activate
@@ -1,1 +1,0 @@
-/home/salman/miniconda3/bin/activate
