--- wa-web/bin/conda
@@ -1,1 +1,0 @@
-/home/salman/miniconda3/bin/conda
