--- wa-web/bin/deactivate
@@ -1,1 +1,0 @@
-/home/salman/miniconda3/bin/deactivate
