--- wa-web/bin/easy_install-2.7
@@ -1,11 +1,0 @@
-#!/home/salman/All-Projects/WhatsApp-Web/wa-web/bin/python2
-
-# -*- coding: utf-8 -*-
-import re
-import sys
-
-from setuptools.command.easy_install import main
-
-if __name__ == '__main__':
-    sys.argv[0] = re.sub(r'(-script\.pyw?|\.exe)?$', '', sys.argv[0])
-    sys.exit(main())
