--- wa-web/bin/pip
@@ -1,11 +1,0 @@
-#!/home/salman/All-Projects/WhatsApp-Web/wa-web/bin/python2
-
-# -*- coding: utf-8 -*-
-import re
-import sys
-
-from pip import main
-
-if __name__ == '__main__':
-    sys.argv[0] = re.sub(r'(-script\.pyw?|\.exe)?$', '', sys.argv[0])
-    sys.exit(main())
