--- wa-web/bin/python
@@ -1,1 +1,0 @@
-python2
