--- wa-web/bin/python2.7
@@ -1,1 +1,0 @@
-python2
