--- wa-web/include/python2.7
@@ -1,1 +1,0 @@
-/usr/include/python2.7
