--- wa-web/lib64
@@ -1,1 +1,0 @@
-lib
