--- wa-web/local/include
@@ -1,1 +1,0 @@
-/home/salman/All-Projects/WhatsApp-Web/wa-web/include
