--- wa-web/local/lib
@@ -1,1 +1,0 @@
-/home/salman/All-Projects/WhatsApp-Web/wa-web/lib
