--- wa-web/local/pyvenv.cfg
@@ -1,1 +1,0 @@
-/home/salman/All-Projects/WhatsApp-Web/wa-web/pyvenv.cfg
