--- wa-web/pyvenv.cfg
@@ -1,3 +1,0 @@
-home = /home/salman/miniconda3/bin
-include-system-site-packages = false
-version = 3.6.0
